# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Запись на спортивные курсы МГТУ",
        "contact": {},
        "version": "0.0-0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/enroll": {
            "put": {
                "description": "Создаёт новую заявку и связывает её с группой/ами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Запись"
                ],
                "summary": "Записать в группу/ы",
                "parameters": [
                    {
                        "description": "Параметры записи",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.EnrollRequestBody"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/enrollment/delete/{enrollment_id}": {
            "put": {
                "description": "Изменяет статус записи на \"Удалён\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Записи"
                ],
                "summary": "Удалить запись",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id записи",
                        "name": "enrollment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/enrollment/edit": {
            "put": {
                "description": "Находит запись и редактирует её поля",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Записи"
                ],
                "summary": "Редактировать запись",
                "parameters": [
                    {
                        "description": "Запись",
                        "name": "enrollment",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/ds.Enrollment"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/enrollment/status_change": {
            "put": {
                "description": "Получает id записи м-м и новый статус и производит необходимые обновления",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Запись"
                ],
                "summary": "Редактировать статус м-м",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.ChangeEnrollmentToGroupStatusRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/enrollment_to_group/delete": {
            "put": {
                "description": "Удаляет запись в таблице enrollment_to_group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enrollments"
                ],
                "summary": "Удаляет связь группы с записью",
                "parameters": [
                    {
                        "description": "Параметры запроса",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.DeleteEnrollmentToGroupRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/enrollments": {
            "get": {
                "description": "Возвращает список всех доступных записей",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Записи"
                ],
                "summary": "Получить записи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Статус записи",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/group/add": {
            "put": {
                "description": "Создает новую группу с параметрами, описанными в json'е",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Группы"
                ],
                "summary": "Добавляет новую группу в БД",
                "parameters": [
                    {
                        "description": "Характеристики новой группы",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Group"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Группа успешно добавлена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/group/delete/{group_title}": {
            "put": {
                "description": "Находит группу по его названию и изменяет его статус на \"Недоступен\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Группы"
                ],
                "summary": "Удалить группу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название группы",
                        "name": "group_title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/group/delete_restore/{group_title}": {
            "get": {
                "description": "Изменяет статус группы с \"Действует\" на \"Недоступен\" и обратно",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Группы"
                ],
                "summary": "Удалить или восстановить группу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название группы",
                        "name": "group_title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/group/edit": {
            "put": {
                "description": "Находит группу по имени и обновляет перечисленные поля",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Группы"
                ],
                "summary": "Редактировать группу",
                "parameters": [
                    {
                        "description": "Данные редактируемого группы (должны содержать имя группы или его id)",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Group"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/groups": {
            "get": {
                "description": "Возвращает все существующие группы",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Группы"
                ],
                "summary": "Получить все существующие группы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Паттерн названия группы",
                        "name": "name_pattern",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Локация",
                        "name": "location",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Статус группы (Действует/Недействителен)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Проверяет данные для входа и в случае успеха возвращает токен для входа",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Вход в систему",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.loginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.loginResp"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "Добавляет в БД нового пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Зарегистрировать нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.registerReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.registerResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.loginReq": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "ds.ChangeEnrollmentToGroupStatusRequestBody": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "ds.DeleteEnrollmentToGroupRequestBody": {
            "type": "object",
            "properties": {
                "enrollmentID": {
                    "type": "integer"
                },
                "groupID": {
                    "type": "integer"
                }
            }
        },
        "ds.EnrollRequestBody": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "ds.Enrollment": {
            "type": "object",
            "properties": {
                "dateCreated": {
                    "type": "string"
                },
                "dateFinished": {
                    "type": "string"
                },
                "dateProcessed": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moderator": {
                    "$ref": "#/definitions/ds.User"
                },
                "moderatorRefer": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/ds.User"
                },
                "userRefer": {
                    "type": "string"
                }
            }
        },
...
    }
}